echochamber/pyproject.toml
[tool.poetry]
name = "echochamber"
version = "0.1.0"
description = ""
authors = ["Adam Yodinsky <27074934+adamyodinsky@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.2.2"
tiktoken = "^0.3.3"
openai = "^0.27.4"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
